Fix handling of boolean_keywords in phrases
The patch of June 11 avoids endless loop when combining phrases
and boolean search, but searches fails.

This patch distinguishes if a boolean keyword is in a phrase or
not.  When it is in a phrase it isn't treated special (that is,
as a boolean), but as a part of the phrase.