Fuzzy.h:allocate memory for name (fixes a0f3cf7)
Thanks for the clue @tauheed

This gets rid of one of the many remaining compiler warnings.

In the other files, I had to change the other ways "name" gets assigned;
declared "name" as a char[]

I defined NAME_LEN_MAX at the top of Fuzzy.h so we could keep track of
the value better. I doubt it will need to get changed though. As far as
I know at this point, the length won't be longer than the strings shown
in this diff.