configure.ac:add PACKAGE_URL and AC_SYS_LARGEFILE
The change at the top will add PACKAGE_URL to include/config.h (when
it's generated after running `./configure`). It will also put that URL
in a few other spots, which will be seen after autoreconf -ivf is run.

The second change I think is all that's needed now to test for large
file support. It will change the option for big/large file (shown when
`./configure --help` is used)