configure:remove remaining instances of bigfile
This is paired with
https://github.com/solbu/hldig/commit/7e86c6ef05ca12826d7f8c9f1c1da8b9d75e3708#diff-67e997bcfdac55191033d57a16d1408a

Instead of channging the option in the .html docs, I removed those
sections. I don't know how outdated that info is about specific systems.
And if largefiles are a problem for people, they will generally know to
checkout for a --disable-largefile option, which we still have.
(fixes #120)